+++ src/CommandAvailability.cpp
@@ -601,7 +601,7 @@ CommandVisibility GetCommandVisibility(int cmdId, const AppCommandCtx& ctx, Comm
     if (!ctx.isPdf) {
         if (cmdId == CmdPdShowInfo || cmdId == CmdPdfBake || cmdId == CmdPdfCompress || cmdId == CmdPdfDecompress ||
             cmdId == CmdPdfEncrypt || cmdId == CmdPdfDecrypt || cmdId == CmdPdfDeletePages ||
-            cmdId == CmdPdfExtractPages) {
+            cmdId == CmdPdfExtractPages || cmdId == CmdTogglePageBoxes) {
             return CommandVisibility::Hide;
         }
     }